push for midterm build
added elevator script (needs more work)
