+++ HorizonAlpha1/Assets/Prefabs/Progression.asset
@@ -18,17 +18,20 @@ MonoBehaviour:
     - stat: 0
       levels:
       - 50
-      - 100
+      - 99
       - 200
       - 400
       - 700
+    health: []
   - characterClass: 1
     stats:
     - stat: 0
       levels:
       - 40
+    health: []
   - characterClass: 3
     stats:
     - stat: 0
       levels:
       - 30
+    health: []
