+++ HorizonAlpha1/Assets/Scripts/Stats/BaseStats.cs
@@ -2,8 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace RPG.Stats
-{
+
   public class BaseStats : MonoBehaviour
   {
       [Range(1, 99)]
@@ -21,4 +20,3 @@ namespace RPG.Stats
         return 10;
       }
   }
-}
