+++ HorizonAlpha1/Assets/Scripts/Stats/Stats.cs
@@ -1,8 +1,5 @@
-namespace RPG.Stats
-{
+
   enum Stat
   {
     Health,
   }
-
-}
