+++ HorizonAlpha1/Assets/elevatorStart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d71901b1bcdc64f99933d887a659b0ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
