+++ HorizonAlpha1/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.10f1
-m_EditorVersionWithRevision: 2021.3.10f1 (1c7d0df0160b)
+m_EditorVersion: 2021.3.9f1
+m_EditorVersionWithRevision: 2021.3.9f1 (ad3870b89536)
