+++ HorizonAlpha1/UserSettings/EditorUserSettings.asset
@@ -18,19 +18,13 @@ EditorUserSettings:
       value: 50540d535056585d5b0c5d7740770647144e4b792e7c2569752c4e61e3b1676b
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 50540d0454000b5e0e0f0e2347745a444e151b7a2d7877687f7b4b36b2b6326c
+      value: 505251505c53515a090d557343775d49404f4a2f7a2e70347e7c4e32b4e1673b
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 505251505c53515a090d557343775d49404f4a2f7a2e70347e7c4e32b4e1673b
+      value: 0206565601505d0d5a0c55714773084413164979742973327d2a1831b6e26c6c
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 5a08575f5207595a0f5d59741173094444164f7d7d2a23317c7a4465bbe1646d
-      flags: 0
-    RecentlyUsedSceneGuid-7:
-      value: 5552560501575a5d0c5f587647745e4443154b2f797f7760752a4f66b3b76169
-      flags: 0
-    RecentlyUsedSceneGuid-8:
-      value: 0206565601505d0d5a0c55714773084413164979742973327d2a1831b6e26c6c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
